refactor(desktop): split gateway bootstrap hook
