+++ packages/desktop/src/renderer/hooks/useGatewayEventListeners.ts
@@ -1,0 +1,18 @@
+import { useEffect } from 'react';
+import type { GatewayDispatcher } from './useGatewayDispatcherSetup';
+
+export function useGatewayEventListeners(dispatcher: GatewayDispatcher): void {
+  useEffect(() => {
+    const removeEvents = dispatcher.start();
+    const removeStatus = dispatcher.startGatewayStatus();
+    dispatcher.initialize();
+    const removeDebug = window.clawwork.onDebugEvent(() => {});
+
+    return () => {
+      removeEvents();
+      removeStatus();
+      removeDebug();
+      dispatcher.reset();
+    };
+  }, [dispatcher]);
+}
